### 2022-01-21 #4089 @whedon
[PRE REVIEW]: GEMMI: A library for structural biology
**Submitting author:** @wojdyr (<a href="http://orcid.org/0000-0003-3980-4092">Marcin Wojdyr</a>)
**Repository:** https://github.com/project-gemmi/gemmi/
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.1
