### 2022-01-22 #4089 @wojdyr
Regarding reviewer suggestion: the joss_reviewers list has a few people with "structural biology" topic:
joaorodrigues shriyaamittal joaomcteixeira rvhonorato nicksisco1932

If you'd like someone from outside this list, you could try people from here:
https://phenix-online.org/people
They won't have any conflicts of interest as we never collaborated.
