### 2022-01-22 8b7b259
mtz: add FreeRflag as another possible label for free flags
this label is used by XDSCONV
