### 2022-01-24 e359e6e
mtz2cif: ignore _something appended to column names in merged MTZ
MTZ files processed in ccp4i can have column names such as
IMEAN_XDSdataset, SIGIMEAN_XDSdataset, I_XDSdataset(+), etc.
If --spec is not given, ignore the added _part.
