### 2022-01-24 019156e
fix compilation with GCC 4.8
