+++ src/merge.cpp
@@ -53,6 +53,7 @@ const option::Descriptor Usage[] = {
 };
 
 using gemmi::Intensities;
+using gemmi::DataType;
 
 void output_intensity_statistics(const Intensities& intensities) {
   size_t plus_count = 0;
@@ -68,32 +69,32 @@ void output_intensity_statistics(const Intensities& intensities) {
 }
 
 void read_intensities_from_rblocks(Intensities& intensities,
-                                   Intensities::Type itype,
+                                   DataType data_type,
                                    std::vector<gemmi::ReflnBlock>& rblocks,
                                    const char* block_name, bool verbose) {
   for (gemmi::ReflnBlock& rb : rblocks) {
     if (block_name && rb.block.name != block_name)
       continue;
-    rb.use_unmerged(itype == Intensities::Type::Unmerged);
+    rb.use_unmerged(data_type == DataType::Unmerged);
     if (!rb.default_loop)
       continue;
-    if (itype == Intensities::Type::Mean && rb.find_column_index("intensity_meas") < 0) {
+    if (data_type == DataType::Mean && rb.find_column_index("intensity_meas") < 0) {
       if (rb.find_column_index("pdbx_I_plus") < 0)
         gemmi::fail("merged intensities not found");
       std::fprintf(stderr, "No _refln.intensity_meas, using pdbx_I_plus/minus ...\n");
-      itype = Intensities::Type::Anomalous;
+      data_type = DataType::Anomalous;
     }
     if (verbose)
       std::fprintf(stderr, "Reading %s from block %s ...\n",
-                   Intensities::type_str(itype), rb.block.name.c_str());
-    intensities.read_mmcif(rb, itype);
-    if (itype != Intensities::Type::Unmerged)
+                   Intensities::type_str(data_type), rb.block.name.c_str());
+    intensities.read_mmcif(rb, data_type);
+    if (data_type != DataType::Unmerged)
       intensities.take_staraniso_b_from_mmcif(rb.block);
     break;
   }
 }
 
-Intensities read_intensities(Intensities::Type itype, const char* input_path,
+Intensities read_intensities(DataType data_type, const char* input_path,
                              const char* block_name, bool verbose) {
   try {
     Intensities intensities;
@@ -102,16 +103,16 @@ Intensities read_intensities(Intensities::Type itype, const char* input_path,
       if (verbose)
         mtz.warnings = stderr;
       mtz.read_input(gemmi::MaybeGzipped(input_path), /*with_data=*/true);
-      if (itype == Intensities::Type::Unknown)
-        itype = mtz.batches.empty() ? Intensities::Type::Mean : Intensities::Type::Unmerged;
-      if (itype == Intensities::Type::Mean && !mtz.imean_column()) {
+      if (data_type == DataType::Unknown)
+        data_type = mtz.batches.empty() ? DataType::Mean : DataType::Unmerged;
+      if (data_type == DataType::Mean && !mtz.imean_column()) {
         std::fprintf(stderr, "No IMEAN, using I(+) and I(-) ...\n");
         if (!mtz.iplus_column())
           gemmi::fail("I(+) not found");
-        itype = Intensities::Type::Anomalous;
+        data_type = DataType::Anomalous;
       }
-      intensities.read_mtz(mtz, itype);
-      if (itype != Intensities::Type::Unmerged)
+      intensities.read_mtz(mtz, data_type);
+      if (data_type != DataType::Unmerged)
         intensities.take_staraniso_b_from_mtz(mtz);
     } else if (gemmi::giends_with(input_path, ".hkl")) {
       gemmi::XdsAscii xds_ascii;
[…]
