### 2022-01-25 06bb068
check_data_type_under_symmetry() -> DataType::Mean/Anomalous/Unmerged
primarily for checking what's in mmCIF _refln category
