+++ src/cif2mtz.cpp
@@ -84,6 +84,13 @@ void print_block_info(gemmi::ReflnBlock& rb, const gemmi::Mtz& mtz) {
   for (const gemmi::Mtz::Column& col : mtz.columns)
     std::printf(" %s", col.label.c_str());
   std::putchar('\n');
+  if (mtz.is_merged()) {
+    gemmi::DataType type = check_data_type_under_symmetry(gemmi::MtzDataProxy{mtz});
+    if (type == gemmi::DataType::Anomalous)
+      std::printf("  NOTE: this is old-style anomalous data.\n");
+    else if (type == gemmi::DataType::Unmerged)
+      std::printf("  NOTE: this is old-style unmerged data.\n");
+  }
   for (const std::string& d : rb.block.find_values("_diffrn.details"))
     std::printf("  details: %s\n", d.c_str());
 }
