+++ python/mtz.cpp
@@ -252,6 +252,7 @@ void add_mtz(py::module& m) {
     }, py::arg("array"))
     .def("update_reso", &Mtz::update_reso)
     .def("sort", &Mtz::sort)
+    .def("ensure_asu", &Mtz::ensure_asu)
     .def("switch_to_original_hkl", &Mtz::switch_to_original_hkl)
     .def("switch_to_asu_hkl", &Mtz::switch_to_asu_hkl)
     .def("write_to_file", &Mtz::write_to_file, py::arg("path"))
