### 2022-01-26 d497495
change version number to 0.5.2
