+++ .github/workflows/wheels2.yml
@@ -30,7 +30,7 @@ jobs:
     - uses: actions/checkout@v2
 
     - name: Build wheels
-      uses: pypa/cibuildwheel@v2.3.0
+      uses: pypa/cibuildwheel@v2.3.1
       env:
         CIBW_BUILD: ${{ github.event.inputs.cibw_build }}
         CIBW_SKIP: ${{ github.event.inputs.cibw_skip }}
