### 2022-01-26 dde43ed
bessel.hpp: use double instead of float, make tables accessible
