### 2022-01-27 2d7184f
add log_bessel_i0()
