### 2022-01-27 338930d
added log_cosh
