### 2022-02-01 5cdc5dd
WIP: Binner (resolution shells)
