### 2022-02-02 6146f16
CIF tag comparisons should be case insensitive
