### 2022-02-02 78c963d
helper function lower(char) -> char (micro-optimization)
