+++ include/gemmi/util.hpp
@@ -6,7 +6,7 @@
 #define GEMMI_UTIL_HPP_
 
 #include <algorithm>  // for equal, find, remove_if
-#include <cctype>     // for tolower
+#include <cctype>     // for isspace
 #include <cstring>    // for strncmp
 #include <iterator>   // for begin, end, make_move_iterator
 #include <string>
@@ -30,35 +30,15 @@ inline bool ends_with(const std::string& str, const std::string& suffix) {
   return str.length() >= sl && str.compare(str.length() - sl, sl, suffix) == 0;
 }
 
-// Case-insensitive comparisons. The second arg must be lowercase.
-
-inline bool iequal(const std::string& str, const std::string& low) {
-  return str.length() == low.length() &&
-         std::equal(std::begin(low), std::end(low), str.begin(),
-                    [](char c1, char c2) { return c1 == std::tolower(c2); });
+// can be faster than std::tolower() b/c it takes char not int
+inline char lower(char c) {
+  if (c >= 'A' && c <= 'Z')
+    return c | 0x20;
+  return c;
 }
 
-inline bool iequal_from(const std::string& str, size_t offset, const std::string& low) {
-  return str.length() == low.length() + offset &&
-         std::equal(std::begin(low), std::end(low), str.begin() + offset,
-                    [](char c1, char c2) { return c1 == std::tolower(c2); });
-}
-
-inline bool istarts_with(const std::string& str, const std::string& prefix) {
-  return str.length() >= prefix.length() &&
-         std::equal(std::begin(prefix), std::end(prefix), str.begin(),
-                    [](char c1, char c2) { return c1 == std::tolower(c2); });
-}
-inline bool iends_with(const std::string& str, const std::string& suffix) {
-  size_t sl = suffix.length();
-  return str.length() >= sl &&
-         std::equal(std::begin(suffix), std::end(suffix), str.end() - sl,
-                    [](char c1, char c2) { return c1 == std::tolower(c2); });
-}
-
-inline bool giends_with(const std::string& str, const std::string& suffix) {
-  return iends_with(str, suffix) || iends_with(str, suffix + ".gz");
-}
+// works as expected only for a-zA-Z
+inline char alpha_up(char c) { return c & ~0x20; }
 
 inline std::string to_lower(std::string str) {
   for (char& c : str)
@@ -74,8 +54,35 @@ inline std::string to_upper(std::string str) {
   return str;
 }
 
-// works as expected only for a-zA-Z
-inline char alpha_up(char c) { return c & ~0x20; }
+// Case-insensitive comparisons. The second arg must be lowercase.
+
+inline bool iequal(const std::string& str, const std::string& low) {
+  return str.length() == low.length() &&
+         std::equal(std::begin(low), std::end(low), str.begin(),
+                    [](char c1, char c2) { return c1 == lower(c2); });
+}
+
+inline bool iequal_from(const std::string& str, size_t offset, const std::string& low) {
+  return str.length() == low.length() + offset &&
+         std::equal(std::begin(low), std::end(low), str.begin() + offset,
+                    [](char c1, char c2) { return c1 == lower(c2); });
+}
+
+inline bool istarts_with(const std::string& str, const std::string& prefix) {
+  return str.length() >= prefix.length() &&
+         std::equal(std::begin(prefix), std::end(prefix), str.begin(),
+                    [](char c1, char c2) { return c1 == lower(c2); });
+}
+inline bool iends_with(const std::string& str, const std::string& suffix) {
+  size_t sl = suffix.length();
+  return str.length() >= sl &&
+         std::equal(std::begin(suffix), std::end(suffix), str.end() - sl,
[…]
