+++ include/gemmi/binner.hpp
@@ -21,30 +21,52 @@ struct Binner {
     Refmac,
   };
 
-  int setup_from_1_d2(int nbins, Method method, std::vector<double> dm2) {
+  int setup_from_1_d2(int nbins, Method method, std::vector<double>&& dm2) {
     if (nbins < 2)
       fail("Binner: at least two bins are needed");
     if (dm2.empty())
       fail("Binner: no data");
-    bins.resize(nbins);
-    std::sort(dm2.begin(), dm2.end());
-    min_1_d2 = dm2.front();
-    max_1_d2 = dm2.back();
+    bin_limits.resize(nbins);
+    if (method == Method::EqualCount) {
+      std::sort(dm2.begin(), dm2.end());
+      min_1_d2 = dm2.front();
+      max_1_d2 = dm2.back();
+    } else {
+      min_1_d2 = max_1_d2 = dm2.front();
+      for (double x : dm2) {
+        if (x < min_1_d2)
+          min_1_d2 = x;
+        if (x > max_1_d2)
+          max_1_d2 = x;
+      }
+    }
     if (method == Method::EqualCount) {
       double avg_count = double(dm2.size()) / nbins;
       for (int i = 1; i < nbins; ++i)
-        bins[i-1] = dm2[int(avg_count * i)];
+        bin_limits[i-1] = dm2[int(avg_count * i)];
     } else if (method == Method::Dstar2) {
-      double step = (dm2.back() - dm2.front()) / nbins;
+      double step = (max_1_d2 - min_1_d2) / nbins;
       for (int i = 1; i < nbins; ++i)
-        bins[i-1] = i * step;
-    } else if (method == Method::Dstar3) {
+        bin_limits[i-1] = min_1_d2 + i * step;
     } else if (method == Method::Dstar) {
+      double min_1_d = std::sqrt(min_1_d2);
+      double max_1_d = std::sqrt(max_1_d2);
+      double step = (max_1_d - min_1_d) / nbins;
+      for (int i = 1; i < nbins; ++i)
+        bin_limits[i-1] = sq(min_1_d + i * step);
+    } else if (method == Method::Dstar3) {
+      double min_1_d3 = min_1_d2 * std::sqrt(min_1_d2);
+      double max_1_d3 = max_1_d2 * std::sqrt(max_1_d2);
+      double step = (max_1_d3 - min_1_d3) / nbins;
+      for (int i = 1; i < nbins; ++i)
+        bin_limits[i-1] = sq(std::cbrt(min_1_d3 + i * step));
     } else if (method == Method::LogDstar) {
+      // TODO
     } else if (method == Method::Refmac) {
+      // TODO
     }
-    bins.back() = std::numeric_limits<double>::infinity();
-    return (int) bins.size();
+    bin_limits.back() = std::numeric_limits<double>::infinity();
+    return (int) bin_limits.size();
   }
 
   template<typename DataProxy>
@@ -53,49 +75,51 @@ struct Binner {
     std::vector<double> dm2(proxy.size() / proxy.stride());
     for (size_t i = 0, offset = 0; i < dm2.size(); ++i, offset += proxy.stride())
       dm2[i] = cell.calculate_1_d2(proxy.get_hkl(offset));
-    return setup_from_1_d2(nbins, method, dm2);
+    return setup_from_1_d2(nbins, method, std::move(dm2));
   }
 
   // Generic. Method-specific versions could be faster.
   int get_bin_number(const Miller& hkl) {
-    if (bins.empty())
+    if (bin_limits.empty())
       fail("Binner not set up");
     double inv_d2 = cell.calculate_1_d2(hkl);
-    auto it = std::lower_bound(bins.begin(), bins.end(), inv_d2);
-    // it can't be bins.end() b/c bins.back() is +inf
[…]
