### 2022-02-03 61a40cb
python: added one more Binner.setup() overload
