### 2022-02-04 f59aca1
changes in Binner and tests
