### 2022-02-04 6bde3fa
fix testing without numpy
