+++ include/gemmi/neutron92.hpp
@@ -1,0 +1,74 @@
+// Copyright 2022 Global Phasing Ltd.
+
+// Neutron coherent scattering lengths of the elements,
+// from Neutron News, Vol. 3, No. 3, 1992.
+//
+// We use the same data as cctbx/eltbx/neutron.h, which is based on
+// https://www.ncnr.nist.gov/resources/n-lengths/list.html
+// which in turn is based on Neutron News, Vol. 3, No. 3, 1992, pp. 29-37.
+
+#ifndef GEMMI_NEUTRON92_HPP_
+#define GEMMI_NEUTRON92_HPP_
+
+#include "formfact.hpp"  // for GaussianCoef
+#include "elem.hpp"      // for El
+
+namespace gemmi {
+
+#if defined(__GNUC__) && __GNUC__-0 > 4
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wfloat-conversion"
+#endif
+
+template<class Real>
+struct Neutron92 {
+  using Coef = GaussianCoef<0, 1, Real>;
+  static Real data[121];
+
+  static Real& get_(El el) { return data[static_cast<int>(el)]; }
+  static bool has(El el) { return get_(el) != 0; }
+  static Coef get(El el) { return Coef{{get_(el)}}; }
+  //static Real* get_ptr(El el) { return &get_(el); }
+};
+
+// real part of the bound coherent neutron scattering length (fm)
+template<class Real>
+Real Neutron92<Real>::data[121] = {
+  /*X*/ 0.0,
+  /*H*/ -3.7390, /*He*/ 3.26,
+  /*Li*/ -1.90, /*Be*/ 7.79, /*B*/ 5.30, /*C*/ 6.646,
+  /*N*/ 9.36, /*O*/ 5.803, /*F*/ 5.654, /*Ne*/ 4.566,
+  /*Na*/ 3.63, /*Mg*/ 5.375, /*Al*/ 3.449, /*Si*/ 4.1491,
+  /*P*/ 5.13, /*S*/ 2.847, /*Cl*/ 9.577, /*Ar*/ 1.909,
+  /*K*/ 3.67, /*Ca*/ 4.70, /*Sc*/ 12.29, /*Ti*/ -3.438,
+  /*V*/ -0.3824, /*Cr*/ 3.635, /*Mn*/ -3.73, /*Fe*/ 9.45,
+  /*Co*/ 2.49, /*Ni*/ 10.3, /*Cu*/ 7.718, /*Zn*/ 5.68,
+  /*Ga*/ 7.288, /*Ge*/ 8.185, /*As*/ 6.58, /*Se*/ 7.97,
+  /*Br*/ 6.795, /*Kr*/ 7.81, /*Rb*/ 7.09, /*Sr*/ 7.02,
+  /*Y*/ 7.75, /*Zr*/ 7.16, /*Nb*/ 7.054,
+  /*Mo*/ 6.715, /*Tc*/ 6.8, /*Ru*/ 7.03, /*Rh*/ 5.88, /*Pd*/ 5.91,
+  /*Ag*/ 5.922, /*Cd*/ 4.87, /*In*/ 4.065, /*Sn*/ 6.225,
[…]
