+++ docs/headers.rst
@@ -178,6 +178,10 @@ gemmi/neighbor.hpp
     Cell-linked lists method for atom searching (a.k.a. grid search, binning,
     bucketing, cell technique for neighbor search, etc).
 
+gemmi/neutron92.hpp
+    Neutron coherent scattering lengths of the elements,
+    from Neutron News, Vol. 3, No. 3, 1992.
+
 gemmi/numb.hpp
     Utilities for parsing CIF numbers (the CIF spec calls it 'numb').
 

[…]
