### 2022-02-09 a7cd4a4
bond_restraint_from_connection(): set *_nucleus variables
