### 2022-02-09 02454a7
read_monomer_lib(): added optional argument libin
with the same meaning as LIBIN in Refmac

also
read_monomer_cif() was moved to class MonLib
crdrst has new option --libin=CIF
