### 2022-02-09 e536335
Topo::initialize_refmac_topology(): if LINK/struct_conn has polymer link
don't use a standard link in this place
