+++ include/gemmi/cif2mtz.hpp
@@ -231,110 +231,109 @@ struct CifToMtz {
       hkl_mover.reset(new UnmergedHklMover(mtz.spacegroup));
       tag.replace(tag_offset, std::string::npos, "diffrn_id");
       int sweep_id_index = loop->find_tag(tag);
+      if (sweep_id_index == -1 && verbose)
+        out << "No diffrn_id. Assuming a single sweep.\n";
       tag.replace(tag_offset, std::string::npos, "pdbx_image_id");
       int image_id_index = loop->find_tag(tag);
-      if (sweep_id_index == -1 || image_id_index == -1) {
-        if (verbose)
+      if (verbose) {
+        if (image_id_index == -1)
           out << "No pdbx_image_id, setting BATCH to a dummy value.\n";
-        auto batch = mtz.batches.emplace(mtz.batches.end());
-        batch->number = 1;
-        batch->set_dataset_id(1);
-        batch->set_cell(mtz.cell);
-        // FIXME should we set more properties in BATCH header?
-      } else {
-        if (verbose)
+        else
           out << "  " << tag << " & diffrn_id -> BATCH\n";
-        struct SweepInfo {
-          std::set<int> frame_ids;
-          int offset;
-          float wavelength = 0.f;
-          std::string crystal_id;
-          int dataset_id;
-        };
-        std::map<int, SweepInfo> sweeps;
-        cif::Block& block = const_cast<cif::Block&>(rb.block);
-        cif::Table tab_w0 = block.find("_diffrn.", {"id", "crystal_id"});
-        cif::Table tab_w1 = block.find("_diffrn_radiation.",
-                                       {"diffrn_id", "wavelength_id"});
-        cif::Table tab_w2 = block.find("_diffrn_radiation_wavelength.",
-                                       {"id", "wavelength"});
-        // store sweep and frame numbers corresponding to reflections
-        batch_nums.reserve(loop->length());
-        for (size_t i = 0; i < loop->values.size(); i += loop->tags.size()) {
-          const std::string& sweep_str = loop->values[i + sweep_id_index];
-          int sweep_id = cif::as_int(sweep_str, 0);
-          const std::string& frame_str = loop->values[i + image_id_index];
-          int frame = 1;
-          if (!cif::is_null(frame_str))
-            frame = (int) std::ceil(cif::as_number(frame_str));
-          batch_nums.push_back({sweep_id, frame});
-          if (frame >= 0) {
-            SweepInfo& sweep = sweeps[sweep_id];
-            if (sweep.frame_ids.empty()) {
-              // new sweep was added - set crystal_id and wavelength
-              try {
-                sweep.crystal_id = tab_w0.find_row(sweep_str).str(1);
-              } catch(std::exception&) {
-                sweep.crystal_id = "unknown";
-              }
-              try {
-                const std::string& wave_id = tab_w1.find_row(sweep_str)[1];
-                const std::string& wavelen = tab_w2.find_row(wave_id)[1];
-                sweep.wavelength = (float) cif::as_number(wavelen, 0.);
-              } catch(std::exception&) {}
-            }
-            sweep.frame_ids.insert(frame);
+      }
+      struct SweepInfo {
+        std::set<int> frame_ids;
+        int offset;
+        float wavelength = 0.f;
+        std::string crystal_id = "unknown";
+        int dataset_id;
+      };
+      std::map<int, SweepInfo> sweeps;
+      cif::Block& block = const_cast<cif::Block&>(rb.block);
+      cif::Table tab_w0 = block.find("_diffrn.", {"id", "crystal_id"});
+      cif::Table tab_w1 = block.find("_diffrn_radiation.",
[…]
