### 2022-02-11 d0467d9
ci.yml: pybind11 just removed support for Python 2
