### 2022-02-11 9e7b928
.appveyor.yml: add VS 2022
