+++ .appveyor.yml
@@ -3,6 +3,14 @@
   environment:
     matrix:
       - arch: x64
+        CMAKE_GENERATOR: Visual Studio 17 2022
+        CMAKE_CONFIG: Debug
+        CMAKE_ARGS: -DCMAKE_CXX_STANDARD=17 -DUSE_PYTHON=1 -DPYTHON_EXECUTABLE="C:\\Python310-x64\\python.exe"
+        APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2022
+        PY_PYTHON: 3.10
+        USE_PIP: 1
+    matrix:
+      - arch: x64
         CMAKE_GENERATOR: Visual Studio 16 2019
         CMAKE_CONFIG: RelWithDebInfo
         CMAKE_ARGS: -DCMAKE_CXX_STANDARD=11 -DUSE_PYTHON=1 -DPYTHON_EXECUTABLE="C:\\Python38-x64\\python.exe"
