+++ .appveyor.yml
@@ -9,7 +9,6 @@
         APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2022
         PY_PYTHON: 3.10
         USE_PIP: 1
-    matrix:
       - arch: x64
         CMAKE_GENERATOR: Visual Studio 16 2019
         CMAKE_CONFIG: RelWithDebInfo
