### 2022-02-11 9e24af7
mtz2cif: add FreeRflag (alternative to FreeR_flag) to default spec
