### 2022-02-11 e2f6f31
avoid MSVC warnings
such as
C:\projects\gemmi\include\gemmi/mtz.hpp(1043,1): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
