### 2022-02-11 #4089 @kellyrowland
Hi @nicksisco1932 @jamesrhester 👋

Would you be interested in and available to review this JOSS submission?
