+++ .github/workflows/ci.yml
@@ -7,7 +7,7 @@ jobs:
     strategy:
       matrix:
         include:
-          - os: macos-10.15
+          - os: macos-11
             build: Release
     name: ${{ matrix.os }}
     runs-on: ${{ matrix.os }}
@@ -41,11 +41,26 @@ jobs:
 
     - uses: actions/upload-artifact@v2
       with:
-        name: gemmi-macos10_15
+        name: gemmi-macos11
         path: |
           ${{runner.workspace}}/build/gemmi
           ${{runner.workspace}}/dist/gemmi*.egg
         retention-days: 14
+
+
+  windows2022:
+    name: "Windows 2022"
+    runs-on: windows-2022
+    if: "!contains(github.event.head_commit.message, '[skip ci]')"
+    steps:
+    - uses: actions/checkout@v2
+    - uses: actions/setup-python@v2
+      with:
+        python-version: "3.10"
+    - name: build and test
+      run: |
+        python -m pip install -v -v .
+        python -m unittest discover -v -s tests/
 
 
   ubuntu2004:
