+++ .appveyor.yml
@@ -5,9 +5,9 @@
       - arch: x64
         CMAKE_GENERATOR: Visual Studio 17 2022
         CMAKE_CONFIG: Release
-        CMAKE_ARGS: -DCMAKE_CXX_STANDARD=14 -DUSE_PYTHON=1 -DPYTHON_EXECUTABLE="C:\\Python39-x64\\python.exe"
+        CMAKE_ARGS: -DCMAKE_CXX_STANDARD=14 -DUSE_PYTHON=1 -DPYTHON_EXECUTABLE="C:\\Python310-x64\\python.exe"
         APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2022
-        PY_PYTHON: 3.9
+        PY_PYTHON: 3.10
         USE_PIP: 1
       - arch: x64
         CMAKE_GENERATOR: Visual Studio 16 2019
@@ -55,7 +55,7 @@
   - py -c "import sys; print(sys.version+'\n'+sys.executable)"
   - if defined USE_SETUP_PY (py setup.py install)
   - cmd: '"%VS140COMNTOOLS%\..\..\VC\vcvarsall.bat" %ARCH%'
-  - if defined USE_PIP (py -m pip install --upgrade .)
+  - if defined USE_PIP (py -m pip install --verbose --upgrade .)
 
   test_script:
   - ctest -C "%CMAKE_CONFIG%" --output-on-failure
