+++ .github/workflows/ci.yml
@@ -7,7 +7,7 @@ jobs:
     strategy:
       matrix:
         include:
-          - os: macos-11
+          - os: macos-10.15
             build: Release
     name: ${{ matrix.os }}
     runs-on: ${{ matrix.os }}
@@ -34,14 +34,14 @@ jobs:
       run: cmake --build . --config ${{ matrix.build }} --target check
 
     - name: Python Setup
-      run: python3 -m pip install .
+      run: python3 -m pip install -v .
 
     - name: Python Tests
       run: python3 -m unittest discover -v -s tests/
 
     - uses: actions/upload-artifact@v2
       with:
-        name: gemmi-macos11
+        name: gemmi-macos10_15
         path: |
           ${{runner.workspace}}/build/gemmi
           ${{runner.workspace}}/dist/gemmi*.egg
@@ -59,7 +59,7 @@ jobs:
         python-version: "3.10"
     - name: build and test
       run: |
-        python -m pip install -v -v .
+        python -m pip install -v .
         python -m unittest discover -v -s tests/
 
 
@@ -109,7 +109,7 @@ jobs:
       run: |
         g++ --version
         python3 setup.py sdist
-        python3 -m pip install dist/gemmi-*.tar.gz
+        python3 -m pip install -v dist/gemmi-*.tar.gz
         python3 -m unittest discover -v -s tests/
         cmake -DCMAKE_CXX_STANDARD=14 -DUSE_FORTRAN=1 -DEXTRA_WARNINGS=1 .
         make
