### 2022-02-12 #4089 @jamesrhester
Yes, happy to review this. I have contributed to a previous project of the author but I don't think this rises to the level of conflict of interest as it is not ongoing and not recent.
