### 2022-02-14 5cdaf53
remove unimplemented methods from Binner
