### 2022-02-15 73eff5e
Binner: add get_bin_dmin() and _dmax()
