### 2022-02-16 8557374
python: add functions binmean(), binrfactor(), bincorr()
They work with bin numbers in the same way as numpy.bincount(),
but they calculate mean, Rfactor and correlation.
