+++ python/gemmi.cpp
@@ -8,9 +8,24 @@
 #include "gemmi/third_party/tao/pegtl/parse_error.hpp" // for parse_error
 
 #include "common.h"
+#include <pybind11/stl.h>
 #include <pybind11/numpy.h>    // for vectorize
 
 namespace py = pybind11;
+
+namespace {
+template<typename T> int get_max_bin(const T& bins) {
+  int max_bin = 0;
+  for (int i = 0; i < bins.shape(0); ++i) {
+    if (bins(i) < 0)
+      throw py::value_error("bins argument must have no negative elements");
+    max_bin = std::max(max_bin, bins(i));
+  }
+  if (max_bin > 1000000)
+    throw py::value_error("bin numbers must be smaller than million");
+  return max_bin;
+}
+} // anonymous namespace
 
 void add_misc(py::module& m) {
   py::class_<gemmi::CifWalk>(m, "CifWalk")
@@ -35,6 +50,72 @@ void add_misc(py::module& m) {
         x = std::abs(x);
         return x + std::log1p(std::exp(-2 * x)) - std::log(2);
   }));
+
+  // utilities inspired by numpy.bincount()
+  m.def("binmean", [](py::array_t<int> bins, py::array_t<double> values) {
+      auto bins_ = bins.unchecked<1>();
+      auto values_ = values.unchecked<1>();
+      auto len = bins_.shape(0);
+      if (len != values_.shape(0))
+        throw std::domain_error("arrays have different lengths");
+      int ret_size = get_max_bin(bins_) + 1;
+      py::array_t<double> ret(ret_size);
+      double* retp = (double*) ret.request().ptr;
+      for (int i = 0; i != ret_size; ++i)
+        retp[i] = 0.;
+      std::vector<int> counts(ret_size);
+      for (int i = 0; i != len; ++i)
+        if (!std::isnan(values_(i))) {
+          int n = bins_(i);
+          counts[n]++;
+          retp[n] += values_(i);
+        }
+      for (int i = 0; i != ret_size; ++i)
+        retp[i] /= counts[i];
+      return ret;
+  }, py::arg("nbins"), py::arg("values"));
+
+  m.def("binrfactor", [](py::array_t<int> bins, py::array_t<double> obs,
+                                                py::array_t<double> calc) {
+      auto bins_ = bins.unchecked<1>();
+      auto obs_ = obs.unchecked<1>();
+      auto calc_ = calc.unchecked<1>();
+      auto len = bins_.shape(0);
+      if (len != obs_.shape(0) || len != calc_.shape(0))
+        throw std::domain_error("arrays have different lengths");
+      int ret_size = get_max_bin(bins_) + 1;
+      py::array_t<double> ret(ret_size);
+      double* retp = (double*) ret.request().ptr;
+      for (int i = 0; i != ret_size; ++i)
+        retp[i] = 0.;
+      std::vector<double> sum_fobs(ret_size);
+      for (int i = 0; i != len; ++i)
+        if (!std::isnan(obs_(i)) && !std::isnan(calc_(i))) {
+          int n = bins_(i);
+          retp[n] += std::fabs(obs_(i) - calc_(i));
+          // Fobs should be positive, but just in case use fabs
+          sum_fobs[n] += std::fabs(obs_(i));
+        }
+      for (int i = 0; i != ret_size; ++i)
+        retp[i] /= sum_fobs[i];
+      return ret;
+  }, py::arg("nbins"), py::arg("obs"), py::arg("calc"));
+
+  m.def("bincorr", [](py::array_t<int> bins, py::array_t<double> obs,
+                                             py::array_t<double> calc) {
[…]
