### 2022-02-16 c4449b0
python: vectorize get_ functions in scaling
