### 2022-02-17 a484abe
Binner: rename methods and write docs
