+++ include/gemmi/binner.hpp
@@ -29,7 +29,7 @@ struct Binner {
       cell = *cell_;
     if (!cell.is_crystal())
       fail("Binner: unknown unit cell");
-    bin_limits.resize(nbins);
+    limits.resize(nbins);
     if (method == Method::EqualCount) {
       std::sort(inv_d2.begin(), inv_d2.end());
       min_1_d2 = inv_d2.front();
@@ -47,13 +47,13 @@ struct Binner {
       case Method::EqualCount: {
         double avg_count = double(inv_d2.size()) / nbins;
         for (int i = 1; i < nbins; ++i)
-          bin_limits[i-1] = inv_d2[int(avg_count * i)];
+          limits[i-1] = inv_d2[int(avg_count * i)];
         break;
       }
       case Method::Dstar2: {
         double step = (max_1_d2 - min_1_d2) / nbins;
         for (int i = 1; i < nbins; ++i)
-          bin_limits[i-1] = min_1_d2 + i * step;
+          limits[i-1] = min_1_d2 + i * step;
         break;
       }
       case Method::Dstar: {
@@ -61,7 +61,7 @@ struct Binner {
         double max_1_d = std::sqrt(max_1_d2);
         double step = (max_1_d - min_1_d) / nbins;
         for (int i = 1; i < nbins; ++i)
-          bin_limits[i-1] = sq(min_1_d + i * step);
+          limits[i-1] = sq(min_1_d + i * step);
         break;
       }
       case Method::Dstar3: {
@@ -69,16 +69,17 @@ struct Binner {
         double max_1_d3 = max_1_d2 * std::sqrt(max_1_d2);
         double step = (max_1_d3 - min_1_d3) / nbins;
         for (int i = 1; i < nbins; ++i)
-          bin_limits[i-1] = sq(std::cbrt(min_1_d3 + i * step));
+          limits[i-1] = sq(std::cbrt(min_1_d3 + i * step));
         break;
       }
     }
-    bin_limits.back() = std::numeric_limits<double>::infinity();
-    return (int) bin_limits.size();
+    limits.back() = std::numeric_limits<double>::infinity();
+    return (int) limits.size();
   }
 
   template<typename DataProxy>
-  int setup(int nbins, Method method, const DataProxy& proxy, const UnitCell* cell_) {
+  int setup(int nbins, Method method, const DataProxy& proxy,
+            const UnitCell* cell_=nullptr) {
     cell = cell_ ? *cell_ : proxy.unit_cell();
     std::vector<double> inv_d2(proxy.size() / proxy.stride());
     for (size_t i = 0, offset = 0; i < inv_d2.size(); ++i, offset += proxy.stride())
@@ -87,65 +88,65 @@ struct Binner {
   }
 
   void ensure_limits_are_set() const {
-    if (bin_limits.empty())
+    if (limits.empty())
       fail("Binner not set up");
   }
 
   // Generic. Method-specific versions could be faster.
-  int get_bin_number_from_1_d2(double inv_d2) {
+  int get_bin_from_1_d2(double inv_d2) {
     ensure_limits_are_set();
-    auto it = std::lower_bound(bin_limits.begin(), bin_limits.end(), inv_d2);
-    // it can't be bin_limits.end() b/c bin_limits.back() is +inf
-    return int(it - bin_limits.begin());
+    auto it = std::lower_bound(limits.begin(), limits.end(), inv_d2);
+    // it can't be limits.end() b/c limits.back() is +inf
+    return int(it - limits.begin());
   }
 
-  int get_bin_number(const Miller& hkl) {
+  int get_bin(const Miller& hkl) {
     double inv_d2 = cell.calculate_1_d2(hkl);
[…]
