### 2022-02-17 b1b7f37
to_mmcif: change missing _struct_mon_prot_cis.label_seq_id from '?' to '.'
(#178)
