### 2022-02-17 f2773a2
to_mmcif: write _atom_site_anisotrop.type_symbol (#178)
