### 2022-02-17 b790acf
doctest: fix another problem
