### 2022-02-17 40ee733
NearestImage::symmetry_code(): make it work for neighbors 5+ unit cells apart
