### 2022-02-18 #4089 @kellyrowland
That's right, generally editors wait until at least two reviewers are
confirmed to create the review issue.

Having said that, if you'd like to get started on the review, I think I
could start that issue and seek a second reviewer in parallel.
