### 2022-02-18 f8c0e2a
appveyor: reduce the number of builds
