### 2022-02-18 4f15b4a
when converting ccp4 map to mask (mode 2 to 0) set all non-zero values to 1
