### 2022-02-20 7aeb39b
floodfill: connect also corner points
Previously 6 neighbours were considered, now 26.
This is done to be fully compatible with Refmac.
