### 2022-02-20 c6ecb49
solvent mask: remove islands before shrinking (not after)
To be compatible with Refmac.
