### 2022-02-20 8f8e706
scripts for comparing masks
