+++ examples/maskcheck.py
@@ -6,11 +6,12 @@
 #   cctbx.python -m mmtbx.command_line.mask file.pdb
 #   maskcheck.py mask.ccp4 file.pdb
 
+import os
 import sys
 import numpy
 import gemmi
 
-def maskcheck(mask_path, coor_path):
+def maskcheck(mask_path, coor_path, output_diff_map=None):
     # read mask
     mask = gemmi.read_ccp4_mask(mask_path)
     mask.setup()
@@ -20,20 +21,29 @@ def maskcheck(mask_path, coor_path):
     st = gemmi.read_structure(coor_path)
 
     # check if unit cell and symmetry are the same
-    if grid.unit_cell.parameters != st.cell.parameters:
-        print('Cell parameters differ:',
-              grid.unit_cell.parameters, 'vs', st.cell.parameters)
+    if not grid.unit_cell.approx(st.cell, epsilon=1e-4):
+        print('Cell parameters differ. PDB:', st.cell.parameters)
+        print('                       Mask:', grid.unit_cell.parameters)
     sg = st.find_spacegroup()
     if grid.spacegroup != sg:
         print('Space groups differ:', grid.spacegroup.xhm(), 'vs', sg.xhm())
 
     # prepare a mask with the same size
-    masker = gemmi.SolventMasker(gemmi.AtomicRadiiSet.Cctbx)
+    if 'REFMAC' in os.environ:
+        print('Generating REFMAC-compatible mask ...')
+        masker = gemmi.SolventMasker(gemmi.AtomicRadiiSet.Refmac)
+        masker.island_min_volume = 8 * 2.8**3
+    else:
+        print('Generating CCTBX-compatible mask ...')
+        masker = gemmi.SolventMasker(gemmi.AtomicRadiiSet.Cctbx)
     grid2 = gemmi.Int8Grid()
     grid2.copy_metadata_from(grid)
     masker.put_mask_on_int8_grid(grid2, st[0])
 
     compare_mask_arrays(grid, grid2)
+    #print_nearby_atoms(st, grid, grid2)
+    if output_diff_map:
+        write_diff_map(grid, grid2, output_diff_map)
 
 
 def compare_mask_arrays(grid1, grid2):
@@ -42,15 +52,33 @@ def compare_mask_arrays(grid1, grid2):
     if arr1.shape != arr2.shape:
         sys.exit('Different grid sizes %s and %s. Exiting.' %
                  (arr1.shape, arr2.shape))
-    else:
-        print('Size: %d x %d x %d,' % arr1.shape, 'total', arr1.size, 'points')
+    print('Size: %d x %d x %d,' % arr1.shape, 'total', arr1.size, 'points')
     t = 2 * (arr1 != 0) + (arr2 != 0)
+    print('File-Gemmi Count Fraction')
     for (a, b) in [(0, 0), (1, 1), (0, 1), (1, 0)]:
         n = numpy.count_nonzero(t == 2*a+b)
         print('%d-%d %12d %6.2f%%' % (a, b, n, 100.*n/arr1.size))
 
+# Print nearby atom for each differing point
+def print_nearby_atoms(st, grid1, grid2):
+    ns = gemmi.NeighborSearch(st[0], st.cell, 4).populate()
+    for (p1, p2) in zip(grid1, grid2):
+        if p1.value != p2.value:
+            pos = grid2.point_to_position(p2)
+            mark = ns.find_nearest_atom(pos)
+            cra = mark.to_cra(st[0])
+            print('%d-%d near %s' % (p1.value, p2.value, cra))
+
+def write_diff_map(grid1, grid2, output_diff_map):
[…]
