### 2022-02-21 fefbe90
avoid msvc warnings
