### 2022-02-21 f89957c
python: add property Grid.array
