### 2022-02-23 #4089 @kellyrowland
Hi @dominiquesydow @betteridiot 👋 would either of you be interested in and available to review this JOSS submission?
