### 2022-02-23 5d5acca
gemmi-map: add option --dump
this is the default action, printing info similar to MAPDUMP
