+++ docs/map-help.txt
@@ -5,6 +5,7 @@ Usage:
   -h, --help         Print usage and exit.
   -V, --version      Print version and exit.
   -v, --verbose      Verbose output.
+  -d, --dump         Print a map summary (default action).
   --deltas           Statistics of dx, dy and dz.
   --check-symmetry   Compare the values of symmetric points.
   --write-xyz=FILE   Write transposed map with fast X axis and slow Z.
