### 2022-02-23 2389ead
tools/compare_mask.sh: simplify
