### 2022-02-23 3fc2ec8
Ccp4.setup(): change arguments - now C++ and Python have the same args
