+++ examples/patterson_slice.py
@@ -5,8 +5,7 @@ from matplotlib import pyplot
 import gemmi
 
 # toxd_aupatt.map is generated by $CCP4/examples/unix/runnable/patterson
-ccp4 = gemmi.read_ccp4_map('/tmp/wojdyr/toxd_aupatt.map')
-ccp4.setup()
+ccp4 = gemmi.read_ccp4_map('/tmp/wojdyr/toxd_aupatt.map', setup=True)
 arr = numpy.array(ccp4.grid, copy=False)
 x = numpy.linspace(0, ccp4.grid.unit_cell.a, num=arr.shape[0], endpoint=False)
 y = numpy.linspace(0, ccp4.grid.unit_cell.b, num=arr.shape[1], endpoint=False)
