+++ include/gemmi/read_map.hpp
@@ -33,7 +33,7 @@ Ccp4<float> read_ccp4_map(const std::string& path, bool setup) {
   Ccp4<float> ccp4;
   ccp4.read_ccp4(MaybeGzipped(path));
   if (setup)
-    ccp4.setup(GridSetup::Full, NAN);
+    ccp4.setup(NAN);
   return ccp4;
 }
 
@@ -41,7 +41,7 @@ Ccp4<int8_t> read_ccp4_mask(const std::string& path, bool setup) {
   Ccp4<int8_t> ccp4;
   ccp4.read_ccp4(MaybeGzipped(path));
   if (setup)
-    ccp4.setup(GridSetup::Full, -1);
+    ccp4.setup(-1);
   return ccp4;
 }
 
