### 2022-02-23 #4200 @danielskatz
> @editorialbot set paper as branch

I don't think this is what the JOSS documentation says to do...  It might need to be updated
