### 2022-02-23 #4200 @kellyrowland
This needs to be updated: https://joss.readthedocs.io/en/latest/editorial_bot.html#compiling-papers-from-a-specific-branch
