### 2022-02-23 e206f2a
add tests for Ccp4.setup()
