+++ tests/test_grid.py
@@ -2,12 +2,9 @@
 
 import math
 import unittest
+import binascii
 import gemmi
-from common import full_path
-try:
-    import numpy
-except ImportError:
-    numpy = None
+from common import full_path, get_path_for_tempfile, assert_numpy_equal, numpy
 
 class TestFloatGrid(unittest.TestCase):
     def test_reading(self):
@@ -79,6 +76,61 @@ class TestFloatGrid(unittest.TestCase):
         m.symmetrize_min()
         self.assertEqual(m.sum(), 2 * N * N * N - 2 * 12)
 
+class TestCcp4Map(unittest.TestCase):
+    @unittest.skipIf(numpy is None, "NumPy not installed.")
+    def test_567_map(self):
+        # make a small, contrived map
+        data = numpy.arange(5*6*7, dtype=numpy.float32).reshape((5,6,7))
+        cell = gemmi.UnitCell(150, 132, 140, 90, 90, 90)
+        m = gemmi.Ccp4Map()
+        m.grid = gemmi.FloatGrid(data, cell, gemmi.SpaceGroup('P 1'))
+        m.update_ccp4_header()
+
+        # write, read and compare
+        tmp_path = get_path_for_tempfile(suffix='.ccp4')
+        m.write_ccp4_map(tmp_path)
+        with open(tmp_path, 'rb') as f:
+            self.assertEqual(binascii.crc32(f.read()), 4078044323)
+
+        box = gemmi.FractionalBox()
+        box.minimum = gemmi.Fractional(0.5/5, 1.5/6, 3.5/7)
+        box.maximum = gemmi.Fractional(4.5/5, 2.5/6, 5.5/7)
+        m.set_extent(box)
+        cut_data = data[1:5, 2:3, 4:6]
+        self.assertEqual(cut_data.shape, (4, 1, 2))
+        self.assertTrue(numpy.array_equal(m.grid.array, cut_data))
+
+        # CCP4 MAPMASK generated tests/iota_yzx.ccp4.gz from map m:
+        #   mapmask mapin iota_full.ccp4 mapout iota_yzx.ccp4 << eof
+        #   XYZLIM 0.3 0.9 3.4 3.45 -0.5 -0.3
+        #   AXIS Y Z X
+        #   MODE mapin
+        #   eof
+        yzx_path = full_path('iota_yzx.ccp4.gz')
+        expanded_data = numpy.full(data.shape, float('nan'), dtype=data.dtype)
+        expanded_data[1:5, 2:3, 4:6] = cut_data
+
+        mcut = gemmi.read_ccp4_map(yzx_path, setup=False)
+        self.assertTrue(mcut.axis_positions(), [1, 2, 0])
+        mcut.setup(float('nan'), gemmi.MapSetup.ReorderOnly)
+        self.assertTrue(mcut.axis_positions(), [0, 1, 2])
+        self.assertFalse(mcut.full_cell())
+        self.assertEqual(mcut.grid.axis_order, gemmi.AxisOrder.Unknown)
+        self.assertTrue(numpy.array_equal(mcut.grid.array, cut_data))
+
+        mcut = gemmi.read_ccp4_map(full_path(yzx_path), setup=False)
+        self.assertFalse(mcut.full_cell())
+        self.assertEqual(mcut.grid.axis_order, gemmi.AxisOrder.Unknown)
+        mcut.setup(float('nan'), gemmi.MapSetup.NoSymmetry)
+        self.assertTrue(mcut.full_cell())
+        #self.assertEqual(mcut.grid.axis_order, gemmi.AxisOrder.XYZ)
+        mcut.setup(float('nan'), gemmi.MapSetup.NoSymmetry)
[…]
