### 2022-02-24 a69fb7a
fix tests
binascii.crc32() gave negative value on Python 2.7 32-bit Windows build.
Let's try zlib.crc32().
