### 2022-02-24 2bd4efa
gemmi-map: if symmetry operators are given, parse them
and check if they correspond to space group from field ISPG
