### 2022-02-24 a4ef58a
fix gemmi-map --check-symmetry
