+++ src/map.cpp
@@ -199,12 +199,13 @@ int GEMMI_MAIN(int argc, char **argv) {
         map.setup(NAN, gemmi::MapSetup::ReorderOnly);
         map.write_ccp4_map(p.options[Reorder].arg);
       }
+      double max_err = 0.;
       if (p.options[CheckSym]) {
-        double max_err = map.setup(NAN, gemmi::MapSetup::NoSymmetry);
-        if (max_err != 0.0)
-          std::printf("Max. difference for point images in P1: %g\n", max_err);
         const double eps = 0.01;
-        max_err = 0;
+        // this is equivalent to calling map.setup(Full), but additionally
+        // it prints inconsistent points.
+        map.setup(NAN, gemmi::MapSetup::NoSymmetry);
+        map.grid.axis_order = gemmi::AxisOrder::XYZ;
         map.grid.symmetrize([&](float a, float b) {
             if (a < b || a > b) {
               double diff = std::fabs(a - b);
@@ -215,15 +216,13 @@ int GEMMI_MAIN(int argc, char **argv) {
             }
             return std::isnan(a) ? b : a;
         });
-        if (max_err != 0.0)
-          std::printf("Max. difference in symmetry images: %g\n", max_err);
+        map.grid.calculate_spacing();
+      } else if (p.options[Full] || p.options[Mask]) {
+        max_err = map.setup(NAN, gemmi::MapSetup::Full);
       }
-      if (p.options[Full] || p.options[Mask]) {
-        double err = map.setup(NAN, gemmi::MapSetup::Full);
-        if (err != 0.0)
-          std::fprintf(stderr, "WARNING: different values for equivalent "
-                               "points, max diff: %g\n", err);
-      }
+      if (max_err != 0.0)
+        std::fprintf(stderr, "WARNING: different values for equivalent "
+                             "points, max diff: %g\n", max_err);
       if (p.options[Full]) {
         size_t nn = std::count_if(map.grid.data.begin(), map.grid.data.end(),
                                   [](float x) { return std::isnan(x); });
