### 2022-02-24 eb7ce5b
add a test for bulk solvent masking
