### 2022-02-25 c0dd738
move two functions from Grid to GridMeta
