### 2022-02-25 ae4edbf
find_blobs_by_flood_fill(): add optional arg negate=false
