### 2022-02-25 7e1076b
improve examples/maskcheck.py
