### 2022-03-04 e4ca4e0
add make_chemcomp_with_restraints()
makes ChemComp with restraints that preserve current values of bonds lengths
and angles, simplified equivalent of Refmac's option "newligand continue"
