### 2022-03-09 5c42199
prepare_topology: more detailed message (#192)
* prepare_topology: more detailed message

* missing space
