### 2022-03-09 7510e0c
topo: do not apply NH3 mod to M-peptide
