### 2022-03-09 2af0908
add function seitz_to_op() - #191
