+++ python/sym.cpp
@@ -234,4 +234,5 @@ void add_symmetry(py::module& m) {
   m.def("find_spacegroup_by_ops", &find_spacegroup_by_ops,
         py::arg("group_ops"), py::return_value_policy::reference,
         "Returns space-group with identical operations.");
+  m.def("seitz_to_op", &seitz_to_op);
 }
