### 2022-03-10 f70fe7b
read_monomer_lib(): avoid copying
