### 2022-03-10 be13aca
one more test for seitz_to_op()
