### 2022-03-10 9458baa
better output from rstdiff.py
