### 2022-03-11 #4200 @jamesrhester
@wojdyr I can't find anything that might be interpreted as "community guidelines" in either the `README.md` or the Introduction on readthedocs. I would be satisifed with a couple of headings along the lines of "Contributing to GEMMI" explaining how to contribute, and another header "Support" explaining support options. The former would make sense in the `README.md` and the latter in the introduction to the documentation, I think.
