### 2022-03-11 #4200 @jamesrhester
Another minor comment: I think it is worth mentioning somewhere in the Summary which major platforms the software compiles and runs on, or maybe just insert the word "cross-platform" somewhere.
