### 2022-03-11 f1a132b
add function normalize_grid()
