### 2022-03-12 #4200 @wojdyr
@jamesrhester Thanks a lot for the review.

We indeed don't have any guidelines or suggestions regarding contributing to GEMMI. Default Github's Community Guidelines apply. I think it makes reporting bugs easier – the user doesn't need to read and conform to a given protocol. Contributing code to GEMMI has not been formalized yet. To formalize it we'd need to decide how to handle copyrights (should we have CLA and copyright assignment from contributors?), and this would need to be discussed with all stakeholders in the project. Which would take time. As for the support, some users open issue or discussion on Github, other use email. The main page of the documentation ([here][1]) encourages to ask question by email, but I don't really have a preferred mode of communication.

"cross-platform" – makes sense, I'll add it (I guess I should wait for the second review before updating the paper).

Thanks again for reviewing the paper!

[1]: https://gemmi.readthedocs.io/en/latest/
