### 2022-03-14 e80ff0f
Grid: add checks for empty grid
