+++ include/gemmi/grid.hpp
@@ -302,8 +302,14 @@ struct Grid : GridBase<T> {
                          w >= 0 ? w : w + nw);
   }
 
+  void check_not_empty() const {
+    if (data.empty())
+      fail("grid is empty");
+  }
+
   // Safe but slower.
   size_t index_s(int u, int v, int w) const {
+    check_not_empty();
     return this->index_q(modulo(u, nu), modulo(v, nv), modulo(w, nw));
   }
 
@@ -338,6 +344,7 @@ struct Grid : GridBase<T> {
 
   // https://en.wikipedia.org/wiki/Trilinear_interpolation
   T interpolate_value(double x, double y, double z) const {
+    check_not_empty();
     int u, v, w;
     double xd = grid_modulo(x, nu, &u);
     double yd = grid_modulo(y, nv, &v);
@@ -415,6 +422,7 @@ struct Grid : GridBase<T> {
   }
   void copy_4x4x4(double& x, double& y, double& z,
                   std::array<std::array<std::array<T,4>,4>,4>& copy) const {
+    check_not_empty();
     auto prepare_indices = [this](double& r, int nt, int (&indices)[4]) {
       int t;
       r = this->grid_modulo(r, nt, &t);
