### 2022-03-14 bfaf904
optimize Grid.get_subarray()  (#90)
