+++ include/gemmi/grid.hpp
@@ -446,6 +446,32 @@ struct Grid : GridBase<T> {
           copy[i][j][k] = this->get_value_q(u_indices[i], v_indices[j], w_indices[k]);
   }
 
+  void get_subarray(T* dest, std::array<int,3> start, std::array<int,3> shape) const {
+    check_not_empty();
+    if (this->axis_order != AxisOrder::XYZ)
+      fail("get_subarray() is for Grids in XYZ order");
+    const int u_start0 = modulo(start[0], nu);
+    for (int w = 0; w < shape[2]; w++) {
+      const int w0 = modulo(start[2] + w, nw);
+      for (int v = 0; v < shape[1]; v++) {
+        const int v0 = modulo(start[1] + v, nv);
+        int u_start = u_start0;
+        const T* src0 = &data[this->index_q(u_start, v0, w0)];
+        int len = shape[0];
+        while (len > nu - u_start) {
+          int elem = nu - u_start;
+          std::copy(src0, src0 + elem, dest);
+          src0 -= u_start;
+          dest += elem;
+          len -= elem;
+          u_start = 0;
+        }
+        std::copy(src0, src0 + len, dest);
+        dest += len;
+      }
+    }
+  }
+
   void set_value(int u, int v, int w, T x) { data[index_s(u, v, w)] = x; }
 
   template <bool UsePbc>
