### 2022-03-15 367d3bd
Ccp4::set_extend(): use get_subgrid()
