### 2022-03-15 1aafaaf
add Grid::set_subarray()
a pair for get_subarray() from #90,
update docs
