### 2022-03-15 66619f0
topology: gemmi and refmac remove terminal modifications
