### 2022-03-15 8674330
change naming of new links to be compatible with Refmac
