### 2022-03-15 6844ccd
python binrfactor(): add riso=True for R_iso calculation
