### 2022-03-17 42992a0
fix test of Ccp4.axis_positions() - #197
