### 2022-03-17 683a3fb
mmcif: fixed reading of _struct_conn.pdbx_dist_value
