### 2022-03-18 1a39f9b
store _chem_comp.type from CCD as ChemComp::group
