### 2022-03-18 91b5407
add helper function ChemComp::has_atom()
