### 2022-03-18 76700c8
python: add ChemComp.find_atom()
