### 2022-03-18 0035f61
add helper function ChemComp::Bond::other()
and use it instead of for_each_bonded_atom()
