+++ include/gemmi/cif2mtz.hpp
@@ -21,11 +21,6 @@ namespace gemmi {
 
 template<typename DataProxy>
 DataType check_data_type_under_symmetry(const DataProxy& proxy) {
-  struct MillerHash {
-    std::size_t operator()(const Miller& hkl) const noexcept {
-      return std::size_t((hkl[0] * 1024 + hkl[1]) * 1024 + hkl[2]);
-    }
-  };
   const SpaceGroup* sg = proxy.spacegroup();
   if (!sg)
     return DataType::Unknown;
