### 2022-03-22 45770db
std::array has defined operator<, no need for std::tie
