### 2022-03-23 b51dc3c
add function combine_correlations()
