### 2022-03-24 8928d1a
python: add operator overloads
