### 2022-03-24 7ff6f9c
expand_pdb_code_to_path(): add option to return SF-mmCIF path
