### 2022-03-24 d10840e
mmCIF: read/write _audit_author.name
