+++ tests/test_mol.py
@@ -181,7 +181,8 @@ class TestMol(unittest.TestCase):
                       if len(output_block.find_mmcif_category(name)) > 0]
         common_categories = [name for name in cnames_out if name in cnames]
         common_categories.sort()
-        cc = ['_atom_site.', '_atom_type.', '_cell.', '_chem_comp.',
+        cc = ['_atom_site.', '_atom_type.', '_audit_author.',
+              '_cell.', '_chem_comp.',
               '_diffrn.', '_diffrn_detector.', '_diffrn_radiation.',
               '_diffrn_source.',
               '_entity.', '_entity_poly_seq.', '_entry.',
