+++ include/gemmi/metadata.hpp
@@ -144,6 +144,7 @@ struct RefinementInfo : BasicRefinementInfo {
 
 
 struct Metadata {
+  std::vector<std::string> authors;  // _audit_author.name
   std::vector<ExperimentInfo> experiments;
   std::vector<CrystalInfo> crystals;
   std::vector<RefinementInfo> refinement;
