### 2022-03-25 f172391
document expand_if_pdb_code()
