### 2022-03-26 28f4c7a
pdb: parse AUTHOR records
