### 2022-03-26 c213297
rename src/crdrst.cpp to src/prep.cpp, include it in main gemmi prog
