+++ run-tests.sh
@@ -13,7 +13,7 @@ BUILD_DIR="$(pwd)"
 PYTHON=`grep PYTHON_EXECUTABLE:FILEPATH= $BUILD_DIR/CMakeCache.txt | cut -d= -f2`
 
 if [ $# = 0 ] || [ $1 != n ]; then
-    (cd $BUILD_DIR && make -j4 all gemmi-crdrst check)
+    (cd $BUILD_DIR && make -j4 all check)
     ./tools/cmp-size.py build/gemmi build/gemmi.*.so
     ./tools/docs-help.sh
 fi
