+++ src/main.cpp
@@ -23,6 +23,7 @@ int merge_main(int argc, char** argv);
 int mondiff_main(int argc, char** argv);
 int mtz_main(int argc, char** argv);
 int mtz2cif_main(int argc, char** argv);
+int prep_main(int argc, char** argv);
 int reindex_main(int argc, char** argv);
 int residues_main(int argc, char** argv);
 int rmsz_main(int argc, char** argv);
@@ -64,6 +65,7 @@ static SubCmd subcommands[] = {
   CMD(mondiff, "compare two monomer CIF files"),
   CMD(mtz, "print info about MTZ reflection file"),
   CMD(mtz2cif, "convert MTZ to structure factor mmCIF"),
+  CMD(prep, "prepare topology file (.crd) for Refmac"),
   CMD(reindex, "reindex MTZ file"),
   CMD(residues, "list residues from a coordinate file"),
   CMD(rmsz, "validate geometry using monomer library"),
