### 2022-03-26 bdae1c6
gemmi-prep: add option --split
