+++ docs/prep-help.txt
@@ -1,13 +1,16 @@
 $ gemmi prep -h
 Usage:
- gemmi prep [options] INPUT_FILE OUTPUT_BASENAME
+ gemmi prep [options] INPUT_FILE OUTPUT_FILE
+ gemmi prep --split [options] INPUT_FILE OUTPUT_BASENAME
 
-Make intermediate files from one of PDB, mmCIF or mmJSON formats.
+Prepare intermediate Refmac files.
+INPUT_FILE can be in PDB, mmCIF or mmJSON format.
 
 Options:
   -h, --help          Print usage and exit.
   -V, --version       Print version and exit.
   -v, --verbose       Verbose output.
+  --split             Split output into two files: crd and rst.
   --monomers=DIR      Monomer library dir (default: $CLIBD_MON).
   --libin=CIF         Custom additions to the monomer library.
   -H, --no-hydrogens  Remove or do not add hydrogens.
