### 2022-03-28 #4200 @jamesrhester
Perhaps then a header "Contributing" with just a "See Github Community Guidelines" link underneath. "Support" could have "Please open an issue (link) or contact via email". Just something minimal like this should be enough.
