### 2022-03-29 8d236dc
python: make ChemComp.name/group mutable
