### 2022-03-29 e5fc518
docs: elaborate on empty cif.Table
