### 2022-03-29 ed1766a
find_blobs_by_flood_fill(): sort the results
