+++ README.md
@@ -25,5 +25,9 @@ GEMMI is an open-source project of [CCP4](https://www.ccp4.ac.uk/)
 and [Global Phasing Ltd](https://www.globalphasing.com/),
 two major providers of software for macromolecular crystallography.
 
+To report problems or ask questions either use GitHub Issues,
+or email wojdyr+gemmi@gmail.com. If you consider a major code contribution,
+please get in touch beforehand.
+
 License: MPLv2, or (at your option) LGPLv3.
 © 2017-2021 Global Phasing Ltd.
