### 2022-03-31 0f28887
add helper function for Grid: mask_nodes_above_threshold()
