### 2022-03-31 2b09be6
add function flood_fill_from_position() for Grid<float>
