+++ python/grid.cpp
@@ -11,7 +11,7 @@ bool operator>(const std::complex<float>& a, const std::complex<float>& b) {
 }
 
 #include "gemmi/grid.hpp"
-#include "gemmi/floodfill.hpp"  // for flood_fill_from_position
+#include "gemmi/floodfill.hpp"  // for flood_fill_above
 #include "gemmi/solmask.hpp"  // for SolventMasker, mask_points_in_constant_radius
 #include "gemmi/blob.hpp"     // for Blob, find_blobs_by_flood_fill
 #include "gemmi/tostr.hpp"
@@ -264,6 +264,6 @@ void add_grid(py::module& m) {
        py::arg("min_score")=15., py::arg("min_peak")=0., py::arg("negate")=false);
 
   // from floodfill.hpp
-  m.def("flood_fill_from_position", &flood_fill_from_position,
+  m.def("flood_fill_above", &flood_fill_above,
         py::arg("grid"), py::arg("seeds"), py::arg("threshold"), py::arg("negate")=false);
 }
