### 2022-03-31 #4200 @wojdyr
@jamesrhester I misunderstood *community guidelines*. I thought about [Github guidelines][1] which recommend that people are nice and don't offend each other. When the user open a new issue, they see a footnote:
<sub>Remember, contributions to this repository should follow our [GitHub Community Guidelines](https://docs.github.com/articles/github-community-guidelines).</sub>

But I see that the community guidelines in the review process are specifically about guidelines how to seek support, report problems, etc. So I added this to README.md:

> To report problems or ask questions either use GitHub Issues, or email [wojdyr+gemmi@gmail.com](mailto:wojdyr+gemmi@gmail.com). If you consider a major code contribution, please get in touch beforehand.

Thanks for the time you spend on this review.

[1]: https://docs.github.com/en/site-policy/github-terms/github-community-guidelines
