### 2022-04-02 67c967a
Topo::initialize_refmac_topology: change handling of duplicated links
If a polymer link is also given in LINK/struct_conn,
use only one of them. If LINK has explicit name, or if it matches
residue-specific link from the library, use it;
otherwise, LINK is repetition of TRANS/CIS, so ignore LINK.
