+++ tools/crd-diff.py
@@ -14,6 +14,7 @@ COLUMNS = ['record', 'number', 'label', 'period',
 
 Restraint = namedtuple('Restraint', COLUMNS)
 Crd = namedtuple('Crd', ['atoms', 'real_serial'])
+ignore_tors = False
 
 def read_rst(doc):
     result = []
@@ -22,6 +23,8 @@ def read_rst(doc):
             result.append((row[0].lower(), row[2].lower(), []))
         else:
             data = Restraint(*[x.lower() if x != '.' else None for x in row])
+            if ignore_tors and data.record == 'tors':
+                continue
             result[-1][2].append(data)
     return [r for r in result if r[2]]
 
@@ -31,6 +34,9 @@ def read_crd(doc):
                                        'label_comp_id', 'occupancy',
                                        'calc_flag'])
     atoms = [a for a in sites if a[-1] != 'M']
+    len_diff = len(sites) - len(atoms)
+    if len_diff != 0:
+        print('Removed %d atoms with calc_flag=M' % len_diff)
     real_serial = {None: None, '.': '.'}
     for a in atoms:
         real_serial[a[0]] = len(real_serial) - 1
@@ -40,13 +46,11 @@ def get_crd_atom_row(crd, atom_id):
     idx = crd.real_serial[atom_id] - 1
     return crd.atoms[idx]
 
-def crd_atom_name(crd, atom_id):
-    return cif.as_string(get_crd_atom_row(crd, atom_id)[1])
-
 # hydrogen distances in Refmac may not be ideal
-def can_have_wrong_val_obs(crd1, rst1):
+def can_have_wrong_val_obs(crd, rst1):
     def is_hydrogen(atom_id):
-        return crd_atom_name(crd1, atom_id).startswith('H')
+        atom = get_crd_atom_row(crd, atom_id)
+        return cif.as_string(atom[1]).startswith('H')
     if rst1.record == 'bond':
         return is_hydrogen(rst1.atom_id_2)
     elif rst1.record == 'angl':
@@ -62,9 +66,14 @@ def can_have_wrong_val_obs(crd1, rst1):
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument('-v', action='store_true', help='verbose output')
+    parser.add_argument('--no-tors', action='store_true',
+                        help='ignore torsion angles')
     parser.add_argument('file1_crd', metavar='file1.crd')
     parser.add_argument('file2_crd', metavar='file2.crd')
     args = parser.parse_args()
+    if args.no_tors:
+        global ignore_tors
+        ignore_tors = True
 
     doc1 = cif.read(args.file1_crd)
     doc2 = cif.read(args.file2_crd)
@@ -79,10 +88,10 @@ def main():
             print('First difference:')
             print('ATOM %s %s %s %s' % (a1[0], a1.str(1), a1[2], a1[3]))
             print('ATOM %s %s %s %s' % (a2[0], a2.str(1), a2[2], a2[3]))
-            break
+            return
         if float(a1[4]) != float(a2[4]):
-            print('ATOM %s %s %s %s occupancy %s vs %s' % (
-                  a1[0], a1.str(1), a1[2], a1[3], a1[4], a2[4]))
+            print('ATOM %s/%s %s %s %s occupancy %s vs %s' % (
+                  a1[0], a2[0], a1.str(1), a1[2], a1[3], a1[4], a2[4]))
 
     r1 = read_rst(doc1)
     r2 = read_rst(doc2)
@@ -105,21 +114,20 @@ def main():
         abs_diff = abs(float(a) - float(b))
         return abs_diff < eps or (mod360 and abs(abs_diff - 360) < eps)
 
-    def fmt_atom_id(atom_id, n):
[…]
