### 2022-04-03 project-gemmi/gemmi#201 @dominiquesydow
JOSS review on software paper
Hi @wojdyr,

I just read through the manuscript - it is well-written and has a clear and concise summary, very much appreciated.

I have a few comments on the "Statement of need" section, referring to this paper version: https://github.com/openjournals/joss-reviews/issues/4200#issuecomment-1049297149

- [x] Could you please elaborate a bit more on the gap that GEMMI is filling?
  - Examples of overlaps and trade-offs are nicely stated - which choices were made for GEMMI and why?
  - Are there examples of functionalities that the other packages do _not_ have?
- [x] If applicable (going hand-in-hand with the previous point), could you state in which cases a user might want to use GEMMI vs. CCTBX vs. Clipper? As someone who has not used any of these software packages, that would be my first question to deciding on which software to spend my time on.
- [x] Does it make sense to mention coot alongside software that has (at least a subset of similar) functionalities?
- [x] UglyMol is mentioned in the "Summary" as an application that uses GEMMI if I understand correctly.  Would make sense to add it also the list at the end of the "Statement of need" alongside autoBUSTER etc.?
- [x] References: If applicable, could you please add DOIs where missing?

Link to the review: https://github.com/openjournals/joss-reviews/issues/4200
