### 2022-04-03 project-gemmi/gemmi#202 @dominiquesydow
JOSS review on documentation - show case?
Hi @wojdyr,

I have had a look at the documentation - great work! Thanks a lot for the many details and for providing example code snippets.
The following question is for your consideration but by no means a request given the already rich documentation.

As a novice to any package, I always welcome it if the authors provide e.g. a Jupyter notebook with a short demo of the package that shows how to use the package for a quite common task in that field. Do you by any chance have such a notebook that could be added to the repository and linked to e.g. the [Introduction](https://gemmi.readthedocs.io/en/latest/index.html#) as a  "Quick start" or "Show case" paragraph?

Link to the review: https://github.com/openjournals/joss-reviews/issues/4200
