### 2022-04-03 project-gemmi/gemmi#204 @dominiquesydow
JOSS review on documentation - community guidelines
Hi @wojdyr and @jamesrhester,

As part of my review, I'd like to follow up the discussion around the community guidelines that already started here: https://github.com/openjournals/joss-reviews/issues/4200#issuecomment-1080081599

As part of the documentation, GEMMI should have:
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

For me, the README addition that you made [here](https://github.com/openjournals/joss-reviews/issues/4200#issuecomment-1085029015) works nicely. I think this still needs to be added also to the documentation - agreeing with @jamesrhester, a "Contributing" section would make sense (e.g. after the "Installation" section).

Link to the review: https://github.com/openjournals/joss-reviews/issues/4200
