### 2022-04-03 project-gemmi/gemmi#201 @dominiquesydow
Adding to the list above, at the moment the paper describes GEMMI nicely as a toolbox for working with crystallographic data. I think it would be beneficial to add a paragraph that shows examples of how this toolbox could be put together for example use cases (e.g. common tasks in a crystallographer's work life).

This goes very much in hand with what I felt was missing when I went through the documentation and described in this issue: https://github.com/project-gemmi/gemmi/issues/202
