### 2022-04-03 #4200 @dominiquesydow
Update from my side: I reviewed the paper and package, leaving comments in the form of issues at https://github.com/project-gemmi/gemmi/issues

Later today, I will do some more testing on this checklist item:
> Functionality: Have the functional claims of the software been confirmed?

The other checklist items are either ticked off or addressed in one of the issues.
