### 2022-04-04 project-gemmi/gemmi#201 @wojdyr
Hi @dominiquesydow

thanks for taking the time to review this paper.

>     * [ ]  Could you please elaborate a bit more on the gap that GEMMI is filling?
>
>       * Examples of overlaps and trade-offs are nicely stated - which choices were made for GEMMI and why?
>       * Are there examples of functionalities that the other packages do _not_ have?

There are many smaller and larger differences, but I couldn't find a representative example that would do justice to all the compared packages. Cctbx has order of magnitude more man-years of development and it doesn't have major gaps, but for sure there are small features in gemmi and clipper that are not in cctbx.

>     * [ ]  If applicable (going hand-in-hand with the previous point), could you state in which cases a user might want to use GEMMI vs. CCTBX vs. Clipper? As someone who has not used any of these software packages, that would be my first question to deciding on which software to spend my time on.

There is no easy answer to this. If the user needs only a small subset of functionality (say, reading a CIF file) they could compare how it can be in different packages, as well as how easy it's to find documentation and install different libraries.
Heavy users will often need custom features or changes in existing features. So they will need to frequently communicate with the project they use. And because they often work in a group doing method development for crystallography, and this groups is usually invested in one of the libraries, they go with what their colleagues use.

>     * [ ]  Does it make sense to mention coot alongside software that has (at least a subset of similar) functionalities?

Coot is using Clipper for most of the similar functionalities. (Coot development started in York, where Clipper is being developed). But thinking about it, I should mention MMDB (another library used by Coot).

>     * [ ]  UglyMol is mentioned in the "Summary" as an application that uses GEMMI if I understand correctly.  Would make sense to add it also the list at the end of the "Statement of need" alongside autoBUSTER etc.?

The list of projects is the Statement of need is far from complete. Github show that [58 repositories][1] depend on gemmi, so I didn't try to list all of them.

>     * [ ]  References: If applicable, could you please add DOIs where missing?

Yes, I'll update the paper soon.

I'll think about examples that could be added. This would need to be a new section in the paper.

Thanks for the review.

[1]: https://github.com/project-gemmi/gemmi/network/dependents?dependent_type=REPOSITORY&package_id=UGFja2FnZS01MjE0Nzg3MA%3D%3D
