### 2022-04-04 project-gemmi/gemmi#202 @wojdyr
Thanks for this idea. I often come across Jupyter notebooks, but I haven't tried creating them myself. So currently I don't have any. But this is something to keep in mind.
