### 2022-04-05 579dd99
add helper function cat() for string concatation
Doesn't use streams like tostr(). This is experimental for now.
