+++ include/gemmi/tostr.hpp
@@ -18,6 +18,14 @@ void add_to_stream(std::ostringstream& os, T&& value, Args&&... args) {
   os << std::forward<T>(value);
   add_to_stream(os, std::forward<Args>(args)...);
 }
+
+inline void add_to_string(std::string&) {}
+
+template <typename T, typename... Args>
+void add_to_string(std::string& out, const T& value, Args const&... args) {
+  out += value;
+  add_to_string(out, args...);
+}
 } // namespace impl
 
 template<typename T, typename... Args>
@@ -25,6 +33,13 @@ std::string tostr(T&& value, Args&&... args) {
   std::ostringstream os;
   impl::add_to_stream(os, std::forward<T>(value), std::forward<Args>(args)...);
   return os.str();
+}
+
+template <class... Args>
+std::string cat(Args const&... args) {
+  std::string out;
+  impl::add_to_string(out, args...);
+  return out;
 }
 
 } // namespace gemmi
