### 2022-04-05 4c11ccd
add SpaceGroup::pdb_name()
As explained in Phenix newsletter CCN_2011_01.pdf#page=12
the PDB uses its own, non-standard symbols for rhombohedral space groups.
