### 2022-04-05 6277f3c
mtz2cif: remove_appendix_from_column_names() printed confusing message
