### 2022-04-07 390f172
prep: write BNDS records
