+++ src/prep.cpp
@@ -262,33 +262,47 @@ cif::Block make_crd(const gemmi::Structure& st, const Topo& topo) {
 }
 
 void add_restraints(const Topo::Rule rule, const Topo& topo,
-                    cif::Loop& restr_loop, int (&counters)[5]) {
+                    cif::Loop& restr_loop, int (&counters)[6],
+                    const gemmi::UnitCell* cell=nullptr) {
+  using namespace gemmi;
   constexpr bool ignore_zero_occ = true;
-  //using gemmi::to_str;
-  const auto& to_str = gemmi::to_str_prec<3>; // to make comparisons easier
-  const auto& to_str3 = gemmi::to_str_prec<3>;
+  const auto& to_str = to_str_prec<3>; // to make comparisons easier
+  const auto& to_str3 = to_str_prec<3>;
   auto to_str_dot = [&](double x) { return std::isnan(x) ? "." : to_str(x); };
   if (rule.rkind == Topo::RKind::Bond) {
     const Topo::Bond& t = topo.bonds[rule.index];
     if (ignore_zero_occ && (t.atoms[0]->occ == 0 || t.atoms[1]->occ == 0))
       return;
-    std::string obs = to_str3(t.calculate()) +
-                      " # " + t.atoms[0]->name + " " + t.atoms[1]->name;
-    restr_loop.add_row({"BOND", std::to_string(++counters[0]),
-                        bond_type_to_string(t.restr->type), ".",
+    double obs;
+    std::string label;
+    int counter_idx;
+    if (cell == nullptr) {  // don't use symmetry
+      counter_idx = 0;
+      obs = t.calculate();
+      label = bond_type_to_string(t.restr->type);
+    } else {
+      counter_idx = 5;
+      NearestImage im = cell->find_nearest_image(t.atoms[0]->pos, t.atoms[1]->pos,
+                                                 Asu::Different);
+      obs = std::sqrt(im.dist_sq);
+      label = im.symmetry_code(true);
+    }
+    restr_loop.add_row({cell == nullptr ? "BOND" : "BNDS",
+                        std::to_string(++counters[counter_idx]),
+                        label, ".",
                         std::to_string(t.atoms[0]->serial),
                         std::to_string(t.atoms[1]->serial),
                         ".", ".",
                         to_str(t.restr->value), to_str(t.restr->esd),
                         to_str_dot(t.restr->value_nucleus),
                         to_str_dot(t.restr->esd_nucleus),
-                        obs});
+                        to_str3(obs) + " # " + t.atoms[0]->name + " " + t.atoms[1]->name});
   } else if (rule.rkind == Topo::RKind::Angle) {
     const Topo::Angle& t = topo.angles[rule.index];
     if (ignore_zero_occ &&
         (t.atoms[0]->occ == 0 || t.atoms[1]->occ == 0 || t.atoms[2]->occ == 0))
       return;
-    std::string obs = to_str3(gemmi::deg(t.calculate()));
+    std::string obs = to_str3(deg(t.calculate()));
     obs += " # " + t.atoms[0]->name + " " +
                    t.atoms[1]->name + " " +
                    t.atoms[2]->name;
@@ -304,7 +318,7 @@ void add_restraints(const Topo::Rule rule, const Topo& topo,
     const Topo::Torsion& t = topo.torsions[rule.index];
     if (ignore_zero_occ && (t.atoms[0]->occ == 0 || t.atoms[3]->occ == 0))
       return;
-    std::string obs = to_str3(gemmi::deg(t.calculate()));
+    std::string obs = to_str3(deg(t.calculate()));
     obs += " # " + t.atoms[0]->name + " " + t.atoms[1]->name +
[…]
