### 2022-04-07 067910a
gemmi-prep: refactoring
