### 2022-04-08 fcd5021
move helper cat() to util.hpp, add cat_to()
