### 2022-04-08 89c1366
replace calls to tostr() with calls to cat()
tostr() is left for python bindings only
