### 2022-04-08 40d8281
to_mmcif.hpp: simplify a bit
