### 2022-04-08 f66edc2
symmetry_code(): change the format used in special cases
Special cases meaning those that are currently incorrect in the PDB,
the ones when a link is made with a symmetry mate from 5+ unit cells apart.
Now mmCIF files in the PDB use format such as 4_7510,
which doesn't conforms to the spec and is ambiguous.
In gemmi we will now use in such cases 4_7_5_10. Previously: 4_75(10).
