### 2022-04-09 612527d
MonLib::match_link() - return the most specific link
and also return is-inverted flag
