### 2022-04-09 7d94786
de-duplicate code in prep.cpp and placeh.hpp
