### 2022-04-09 2db8fd9
gemmi-prep: clean up related to zero-occupancy
