### 2022-04-09 2adf97c
move part of gemmi-prep to the library (crd.hpp)
