+++ src/prep.cpp
@@ -1,22 +1,14 @@
 // Copyright 2017-2022 Global Phasing Ltd.
 
-#include <stdio.h>
-#include <cstdlib>   // for getenv
-#include <algorithm> // for count_if
-#include <iostream>  // for cerr
-#include <stdexcept>
-#include "gemmi/chemcomp.hpp"  // for ChemComp
-#include "gemmi/to_cif.hpp"    // for write_cif_to_stream
+#include <stdio.h>             // for printf, fprintf
+#include <cstdlib>             // for getenv
+#include <iostream>            // for cerr
+#include <stdexcept>           // for exception
+#include "gemmi/crd.hpp"       // for prepare_crd, prepare_rst
+#include "gemmi/to_cif.hpp"    // for write_cif_block_to_stream
 #include "gemmi/fstream.hpp"   // for Ofstream
-#include "gemmi/enumstr.hpp"   // for entity_type_to_string
-#include "gemmi/to_mmcif.hpp"  // for write_struct_conn
-#include "gemmi/sprintf.hpp"   // for to_str, to_str_prec
-#include "gemmi/calculate.hpp" // for find_best_plane
-#include "gemmi/modify.hpp"    // for remove_hydrogens
 #include "gemmi/polyheur.hpp"  // for setup_entities
 #include "gemmi/monlib.hpp"    // for MonLib, read_monomer_lib
-#include "gemmi/topo.hpp"      // for Topo
-#include "gemmi/placeh.hpp"    // for place_hydrogens
 #include "gemmi/read_cif.hpp"  // for read_cif_gz
 #include "gemmi/read_coor.hpp" // for read_structure_gz
 
@@ -24,8 +16,6 @@
 #include "options.h"
 
 namespace cif = gemmi::cif;
-using gemmi::Topo;
-using gemmi::Restraints;
 
 namespace {
 
@@ -58,374 +48,6 @@ const option::Descriptor Usage[] = {
   //  "  --no-zero-occ  \tNo restraints for zero-occupancy atoms." },
   { 0, 0, 0, 0, 0, 0 }
 };
-
-
-bool has_anisou(const gemmi::Model& model) {
-  for (const gemmi::Chain& chain : model.chains)
-    for (const gemmi::Residue& res : chain.residues)
-      for (const gemmi::Atom& a : res.atoms)
-        if (a.aniso.nonzero())
-          return true;
-  return false;
-}
-
-std::string refmac_calc_flag(const gemmi::Atom& a) {
-  switch (a.calc_flag) {
-    case gemmi::CalcFlag::NotSet: return ".";
-    // Refmac seems to be using only flags R and M
-    case gemmi::CalcFlag::Calculated: return a.is_hydrogen() ? "R" : "M";
-    // I think these are not used
-    case gemmi::CalcFlag::Determined: return "d";
-    case gemmi::CalcFlag::Dummy: return "dum";
-  }
-  gemmi::unreachable();
-}
-
-// Get value of _entity_poly_seq.ccp4_mod_id compatible with makecif.
-std::string get_ccp4_mod_id(const std::vector<std::string>& mods) {
-  for (const std::string& m : mods)
-    if (!gemmi::starts_with(m, "DEL-OXT") &&
-        !gemmi::starts_with(m, "DEL-HN") && m != "DEL-NMH")
-      return m;
-  return ".";
-}
-
-cif::Block make_crd(const gemmi::Structure& st, const Topo& topo) {
-  using gemmi::to_str;
-  auto e_id = st.info.find("_entry.id");
-  std::string id = cif::quote(e_id != st.info.end() ? e_id->second : st.name);
-  cif::Block block("structure_" + id);
-  auto& items = block.items;
-
[…]
