+++ include/gemmi/crd.hpp
@@ -1,0 +1,384 @@
+// Copyright 2022 Global Phasing Ltd.
+//
+// Generate Refmac intermediate (prepared) files crd and rst
+
+#ifndef GEMMI_CRD_HPP_
+#define GEMMI_CRD_HPP_
+
+#include <cmath>         // for sqrt
+#include "topo.hpp"      // for Topo
+#include "placeh.hpp"    // for prepare_topology
+#include "enumstr.hpp"   // for entity_type_to_string
+#include "to_mmcif.hpp"  // for write_struct_conn
+#include "sprintf.hpp"   // for to_str, to_str_prec
+#include "calculate.hpp" // for find_best_plane
+
+namespace gemmi {
+
+inline bool has_anisou(const Model& model) {
+  for (const Chain& chain : model.chains)
+    for (const Residue& res : chain.residues)
+      for (const Atom& a : res.atoms)
+        if (a.aniso.nonzero())
+          return true;
+  return false;
+}
+
+inline std::string refmac_calc_flag(const Atom& a) {
+  switch (a.calc_flag) {
+    case CalcFlag::NotSet: return ".";
+    // Refmac seems to be using only flags R and M
+    case CalcFlag::Calculated: return a.is_hydrogen() ? "R" : "M";
+    // I think these are not used
+    case CalcFlag::Determined: return "d";
+    case CalcFlag::Dummy: return "dum";
+  }
+  unreachable();
+}
+
+// Get value of _entity_poly_seq.ccp4_mod_id compatible with makecif.
+inline std::string get_ccp4_mod_id(const std::vector<std::string>& mods) {
+  for (const std::string& m : mods)
+    if (!starts_with(m, "DEL-OXT") &&
+        !starts_with(m, "DEL-HN") &&
+        m != "DEL-NMH")
+      return m;
+  return ".";
+}
+
+inline cif::Block prepare_crd(const Structure& st, const Topo& topo) {
+  auto e_id = st.info.find("_entry.id");
+  std::string id = cif::quote(e_id != st.info.end() ? e_id->second : st.name);
+  cif::Block block("structure_" + id);
+  auto& items = block.items;
+
+  items.emplace_back("_entry.id", id);
+  items.emplace_back("_database_2.code_PDB", id);
+  auto keywords = st.info.find("_struct_keywords.pdbx_keywords");
+  if (keywords != st.info.end())
+    items.emplace_back("_struct_keywords.text", cif::quote(keywords->second));
+  auto title = st.info.find("_struct.title");
+  if (title != st.info.end())
+    items.emplace_back(title->first, cif::quote(title->second));
+  auto initial_date =
+         st.info.find("_pdbx_database_status.recvd_initial_deposition_date");
+  if (initial_date != st.info.end())
+    items.emplace_back("_audit.creation_date", initial_date->second);
+  items.emplace_back("_software.name", "gemmi");
+
+  items.emplace_back(cif::CommentArg{"############\n"
+                                     "## ENTITY ##\n"
+                                     "############"});
+  cif::Loop& entity_loop = block.init_mmcif_loop("_entity.", {"id", "type"});
+  for (const Entity& ent : st.entities)
+    entity_loop.add_row({ent.name, entity_type_to_string(ent.entity_type)});
+  items.emplace_back(cif::CommentArg{"#####################\n"
+                                     "## ENTITY_POLY_SEQ ##\n"
+                                     "#####################"});
+  cif::Loop& poly_loop = block.init_mmcif_loop("_entity_poly_seq.", {
+              "mon_id", "ccp4_auth_seq_id", "entity_id",
[…]
