### 2022-04-09 5f8aad1
initialize_refmac_topology(): adjust for link naming change in refmac
