### 2022-04-09 a9c54f7
tweak crd-diff.py
