### 2022-04-11 cfffc52
add cif::Style::Aligned - writing left-aligned columns in loops
