+++ include/gemmi/to_cif.hpp
@@ -17,6 +17,7 @@ enum class Style {
   PreferPairs,  // write single-row loops as pairs
   Pdbx,         // PreferPairs + put '#' (empty comments) between categories
   Indent35,     // start values in pairs from 35th column
+  Aligned,      // columns in tables are left-aligned
 };
 
 // CIF files are read in binary mode. It makes difference only for text fields.
@@ -40,7 +41,7 @@ inline void write_out_pair(std::ostream& os, const std::string& name,
   } else {
     if (name.size() + value.size() > 120)
       os.put('\n');
-    else if (style == Style::Indent35 && name.size() < 34)
+    else if ((style == Style::Indent35 || style == Style::Aligned) && name.size() < 34)
       os.write("                                  ", 34 - name.size());
     else
       os.put(' ');
@@ -49,22 +50,8 @@ inline void write_out_pair(std::ostream& os, const std::string& name,
   os.put('\n');
 }
 
-inline void write_out_loop_values(std::ostream& os, const Loop& loop) {
-  size_t ncol = loop.tags.size();
-  size_t col = 0;
-  for (const std::string& val : loop.values) {
-    bool text_field = is_text_field(val);
-    os.put(col++ == 0 || text_field ? '\n' : ' ');
-    if (text_field)
-      write_text_field(os, val);
-    else
-      os << val;
-    if (col == ncol)
-      col = 0;
-  }
-}
-
 inline void write_out_loop(std::ostream& os, const Loop& loop, Style style) {
+  constexpr size_t max_padding = 30;  // if increased, adjust os.write() below
   if (loop.values.empty())
     return;
   if ((style == Style::PreferPairs || style == Style::Pdbx) &&
@@ -73,10 +60,42 @@ inline void write_out_loop(std::ostream& os, const Loop& loop, Style style) {
       write_out_pair(os, loop.tags[i], loop.values[i], style);
     return;
   }
+  // tags
   os << "loop_";
   for (const std::string& tag : loop.tags)
     os << '\n' << tag;
-  write_out_loop_values(os, loop);
+  // values
+  size_t ncol = loop.tags.size();
+
+  std::vector<size_t> col_width;
+  if (style == Style::Aligned) {
+    col_width.resize(ncol, 1);
+    size_t col = 0;
+    for (const std::string& val : loop.values) {
+      col_width[col] = std::max(col_width[col], val.size());
+      if (++col == ncol)
+        col = 0;
+    }
+    for (size_t& w : col_width)
+      w = std::min(w, max_padding);
+  }
+
+  size_t col = 0;
+  for (const std::string& val : loop.values) {
+    bool text_field = is_text_field(val);
+    os.put(col == 0 || text_field ? '\n' : ' ');
+    if (text_field)
+      write_text_field(os, val);
+    else
+      os << val;
+    if (col != ncol - 1) {
+      if (!col_width.empty() && val.size() < col_width[col])
+        os.write("                                  ", col_width[col] - val.size());
+      ++col;
+    } else {
+      col = 0;
+    }
+  }
   os.put('\n');
 }
 
