### 2022-04-11 dfaa759
python: add binding of HklMatch.pos
