### 2022-04-12 c67fbd4
DensityCalculator: make it work without setting d_min
Initially, the grid was meant to be initiallized based on d_min.
Now it's possible to use custom grid set up manually by the user.
