### 2022-04-12 15707b8
as_refln_blocks(): if unit cell is missing in block, use first unit cell
for #205
