+++ docs/gemmi-help.txt
@@ -1,5 +1,5 @@
 $ gemmi -h
-gemmi 0.5.3
+gemmi 0.5.4
 Command-line utility that accompanies the GEMMI library,
 which is a joint project of CCP4 and Global Phasing Ltd.
 Licence: Mozilla Public License 2.0.
