### 2022-04-13 7c096f0
scaling: skip nan
