### 2022-04-13 ca803c8
python bindings to SMat33::added_kI()
