### 2022-04-13 924c99a
mtz2cif: clarify combination of --depo and --spec
