### 2022-04-19 0b2379b
fix mtz2cif: status o or f was used when x was meant
