+++ tests/test_prog.py
@@ -107,7 +107,7 @@ S--unday
 $ gemmi mtz2cif tests/5e5z.mtz -
 [...] -158 -155
 1 2 0 46.859 0.6285 68.443 0.4591 o
-1 2 1 ? ? ? ? o
+1 2 1 ? ? ? ? x
 1 2 2 38.249 0.3855 61.841 0.3117 f
 ''')
 
