### 2022-04-19 4b87982
gemmi-mtz2cif: add option --skip-negative-sigi
