+++ docs/cif2mtz-help.txt
@@ -17,6 +17,7 @@ Options:
   -H LINE, --history=LINE  Add a history line.
   -u, --unmerged           Write unmerged MTZ file(s).
   --sort                   Order reflections according to Miller indices.
+  --skip-negative-sigma    Skip reflections with sigma<0 (in any Q column).
   --local                  Take file from local copy of the PDB archive in
                            $PDB_DIR/structures/divided/structure_factors/
 
