### 2022-04-20 0c4225b
python: add Structure.__setitem__
