### 2022-04-20 4a25fc1
Mark::dist_sq() is for internal use
