### 2022-04-20 8cc9e25
fix merge_atoms_in_expanded_model()
and simplify a bit - occupances of all images must be the same
