### 2022-04-20 c486251
gemmi-convert --asembly: fix atom merging
