### 2022-04-22 95f9ac2
more tests
3wup.json.gz was downloaded from PDBe
