### 2022-04-22 749396a
Ccp4::setup(): avoid calling symmetrize() when not necessary
