+++ paper.bib
@@ -85,7 +85,9 @@ volume={34},
 number={18},
 pages={3094--3100},
 year={2018},
-publisher={Oxford University Press}
+publisher={Oxford University Press},
+doi = {10.1093/bioinformatics/bty191},
+url = {https://doi.org/10.1093/bioinformatics/bty191}
 }
 @article{qcp,
 title={Fast determination of the optimal rotational matrix for macromolecular superpositions},
@@ -95,7 +97,9 @@ volume={31},
 number={7},
 pages={1561--1563},
 year={2010},
-publisher={Wiley Online Library}
+publisher={Wiley Online Library},
+doi = {10.1002/jcc.21439},
+url = {https://doi.org/10.1002/jcc.21439}
 }
 @inproceedings{larch,
 title={Larch: an analysis package for XAFS and related spectroscopies},
@@ -105,7 +109,9 @@ volume={430},
 number={1},
 pages={012007},
 year={2013},
-organization={IOP Publishing}
+organization={IOP Publishing},
+doi = {10.1088/1742-6596/430/1/012007},
+url = {https://doi.org/10.1088/1742-6596/430/1/012007}
 }
 @article{cctbx,
 title={The {C}omputational {C}rystallography {T}oolbox: {c}rystallographic algorithms in a reusable software framework},
@@ -115,7 +121,9 @@ volume={35},
 number={1},
 pages={126--136},
 year={2002},
-publisher={International Union of Crystallography}
+publisher={International Union of Crystallography},
+doi = {10.1107/S0021889801017824},
+url = {https://doi.org/10.1107/S0021889801017824}
 }
 @article{clipper,
 title={The {C}lipper {C}++ libraries for {X}-ray crystallography},
