### 2022-04-24 project-gemmi/gemmi#204 @wojdyr
I just edited the manual (page [index.html](https://gemmi.readthedocs.io/en/latest/index.html)) to mention GitHub Issues and Discussions.
It's not worded in the same way as in the README, but please consider the README as part of the documentation (the "Contributing" section being in README is already in the documentation).
