### 2022-04-24 project-gemmi/gemmi#201 @wojdyr
Thanks for thoughtful reading the paper. Comparison with other packages is a sensitive thing and it would be difficult to make for me, because I'm obviously biased. I'll rather wait for someone else to write it. But I understand such comparisons are useful. Recently I was trying to make sense of the landscape in molecular dynamics packages, but I haven't found any comparisons in documentations of these packages.

I added DOIs to all references (that have DOIs) and updated the paper.
