### 2022-04-24 #4200 @wojdyr
@jamesrhester and @dominiquesydow
I appreciate your reading and reviewing of the paper.
I updated the paper, adding words "cross-platform" (as suggested by James) and missing DOIs.
