### 2022-04-26 21ff3bf
scaling: fix fit_isotropic_b_approximately
