### 2022-04-27 #4200 @jamesrhester
Thanks for those updates. One final request from me: I see that the source distribution has a file `run-tests.sh` with an explanation inside on how to run the tests. To make this easier to find (and in order to tick off the checklist item on automated tests) I suggest that a short `Testing` section is added in the 'Installation' section of the docs pointing to this file with a brief instruction on how to run it e.g. "`run-tests.sh` in the distribution will execute a wide variety of tests when executed at the bash prompt". This is the last item on my review checklist.
