### 2022-04-28 #4200 @kellyrowland
@wojdyr I'm ready to recommend this for acceptance. Please create an archive of the tagged release you wish to have the paper associated with (I see we started at v0.5.1 but the software is up to v0.5.4) and archive it (on Zenodo, figshare, or elsewhere) and post the DOI here.
