### 2022-04-28 #4200 @wojdyr
Thank you all!
I uploaded v0.5.1 here:
https://doi.org/10.6084/m9.figshare.19678776
