### 2022-04-29 0bdabbd
Niggli reduction: optionally, calculate the change-of-basis transformation
