### 2022-04-29 03efda6
SellingVector: add more ctors
