### 2022-05-02 38e0a4a
python: made SmallStructure::Site vector opaque
